lib: add libopencm3 subproject
Ideally this would provide a complete meson build system, but libopencm3 includes some complexities (code generation) that make it not trivial to do so.

To save myself from madness I opted to integrate it with the external project meson module instead (unstable) and hack it a bit to make it work.
Why 'hack'? The libopencm3 build system is not very flexible, it does not provide a way to build out-of-tree, and does not provide an install target,
both are required for the meson external project module to work, the hack is a script copying the required sources and a wrapper Makefile.